Add files via upload
Open OOP1_Animal.1vproj, the frame of project is viewed.
Example.vi is the test project. We can change the "Combine box" to choose "dog" or "cat". When the value of Combine box changes, computer will send a messege box to show "小猫喵喵叫" or "小狗汪汪叫".
There are 3 classes. "animal.lvclass" is the father class. "cat.lvclass" and "dog.lvclass" are the son class.